Removed cast from Setting to const char*.
To address hyperrealm/libconfig#1, the cast operator from Setting to
const char* has been removed. As a result, lookupValue() taking a const
char* has also been removed as it uses the cast operator.

This change allows code that assigns directly to a std::string or a
const char* to compile without errors. No change to existing code should
be required.