libconfig: Allow user to select output format types
According to BNF grammar each line traling ; is optional,
so user should be alowed to select if ; shold be placed
in output of library or not.

Similar situation is with : and =. In grammar they are
considered to be equivalent but when using library to export
the config there is some policy which determines which one
should be used. In this situation user should be able to
enforce than one of them should be used in all cases.

Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>