Recognize long integers without the need for L extension
A value is recognized as CONFIG_TYPE_INT if it's only numbers, and
CONFIG_TYPE_INT64 if it's numbers followed by L or LL. However from a
user point of view adding L or LL is confusing. Instead, if the
integer is too big to fit in an int, return a CONFIG_TYPE_INT64. This
also avoid silent truncation.