Add support for binary integer values
Modifiy the libconfig grammar to include binary integers (and 64-bit integers),
prefixed with 0b (or 0B). This also adds a binary integer format so that
they can then be output as binary.

Binaries are output using a custom formatter, even though some C
libraries now support %b/%B format specifiers. This could be improved in
the future.

The code is based heavily on the hexadecimal handling (so, e.g. only the
L serves as an indicator that it should be 64-bits, not the number of
zeroes).

tests for binary and hex integers

don't accidentally define YYDEBUG

spacing fixes

   Make all integer parsing consistent

    This commit does a few different things:

    1) Includes error checking for all integer types (integer, hex, bin)
    2) This includes swalling anny errno set by stro[u]ll and restoring
       errno to original tate
    3) Automatically extends hex and bin to 64 bits even with no L
       present. In this case, we only check against INT_MAX since it's
       rare you'd want to consider negative numbers here, I think.

    We nowalways use the parse_integer/hex/bin functions in the scanner,
    which now all fill ok (and eat errno if necessary). They also all take
    an L_ok argument, which determines if it's ok have L's at the end of the
    digits (this is set to 1 for integer64/hex64/bin64, 0 otherwise).

    Note that the previous code was setting errno to 0 on failure rather
    than to err_save, which I think was probably a bug, unless I'm missing
    something

add tests for automatic bin/hex enlargement

first pass at updated docs

typo in version