Add support for binary integer values
Modifiy the libconfig grammar to include binary integers (and 64-bit integers),
prefixed with 0b (or 0B). This also adds a binary integer format so that
they can then be output as binary.

Binaries are output using a custom formatter, even though some C
libraries now support %b/%B format specifiers. This could be improved in
the future.

The code is based heavily on the hexadecimal handling (so, e.g. only the
L serves as an indicator that it should be 64-bits, not the number of
zeroes).